@@ -1,5 +1,7 @@
 module zgo.at/zli
 
-go 1.13
+go 1.18
 
 require golang.org/x/term v0.14.0
+
+require golang.org/x/sys v0.14.0 // indirect
