@@ -7,13 +7,13 @@ import "fmt"
 func EraseLine() { Stdout.Write([]byte("\x1b[2K\r")) }
 
 // ReplaceLine replaces the current line.
-func ReplaceLine(a ...interface{}) {
+func ReplaceLine(a ...any) {
 	EraseLine()
 	fmt.Fprint(Stdout, a...)
 }
 
 // ReplaceLinef replaces the current line.
-func ReplaceLinef(s string, a ...interface{}) {
+func ReplaceLinef(s string, a ...any) {
 	EraseLine()
 	fmt.Fprintf(Stdout, s, a...)
 }
